requirements.txt
apache-airflow
pandas
yfinance
s3fs
minio
scikit-learn
plotly